+++ ee/tabby-ui/app/(dashboard)/reports/components/constants.ts
@@ -1,0 +1,2 @@
+export const KEY_SELECT_ALL = 'all'
+export const DEFAULT_DATE_RANGE = '-14d'
