+++ ee/tabby-ui/app/(dashboard)/reports/components/report.tsx
@@ -3,46 +3,22 @@
 import { useState } from 'react'
 import { useSearchParams } from 'next/navigation'
 import { eachDayOfInterval } from 'date-fns'
-import { sum } from 'lodash-es'
 import moment from 'moment'
-import numeral from 'numeral'
-import { DateRange } from 'react-day-picker'
 import seedrandom from 'seedrandom'
 import { useQuery } from 'urql'
 
 import {
-  DailyStatsInPastYearQuery,
-  DailyStatsQuery,
-  Language
+  ChatDailyStatsInPastYearQuery,
+  DailyStatsInPastYearQuery
 } from '@/lib/gql/generates/graphql'
 import { useAllMembers } from '@/lib/hooks/use-all-members'
 import { useIsDemoMode } from '@/lib/hooks/use-server-info'
-import { getLanguageDisplayName } from '@/lib/language-utils'
-import { queryDailyStats, queryDailyStatsInPastYear } from '@/lib/tabby/query'
-import { cn } from '@/lib/utils'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-  CommandSeparator
-} from '@/components/ui/command'
-import {
-  IconActivity,
-  IconCheck,
-  IconChevronUpDown,
-  IconCode,
-  IconMessageSquare,
-  IconUsers
-} from '@/components/ui/icons'
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger
-} from '@/components/ui/popover'
+  chatDailyStatsInPastYearQuery,
+  dailyStatsInPastYearQuery
+} from '@/lib/tabby/query'
+import { ArrayElementType } from '@/lib/types'
+import { IconUsers } from '@/components/ui/icons'
 import {
   Select,
   SelectContent,
@@ -52,148 +28,56 @@ import {
   SelectValue
 } from '@/components/ui/select'
 import { Skeleton } from '@/components/ui/skeleton'
-import DateRangePicker from '@/components/date-range-picker'
 import LoadingWrapper from '@/components/loading-wrapper'
 import { SubHeader } from '@/components/sub-header'
 
 import { AnnualActivity } from './annual-activity'
 import { ChatDailyActivity } from './chat-daily-activity'
-import { DailyActivity } from './completion-daily-activity'
-
-const KEY_SELECT_ALL = 'all'
-const DEFAULT_DATE_RANGE = '-14d'
-
-function StatsSummary({
-  dailyStats
-}: {
-  dailyStats?: DailyStatsQuery['dailyStats']
-}) {
-  const totalViews = sum(dailyStats?.map(stats => stats.views))
-  const totalAcceptances = sum(dailyStats?.map(stats => stats.selects))
-  const acceptRate =
-    totalAcceptances === 0
-      ? 0
-      : ((totalAcceptances / totalViews) * 100).toFixed(2)
-  return (
-    <div className="flex w-full flex-col items-start justify-center space-y-3 md:flex-row md:items-center md:space-x-6 md:space-y-0 xl:justify-start">
-      <Card className="flex flex-1 flex-col justify-between self-stretch bg-primary-foreground/30 lg:block">
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">Acceptance Rate</CardTitle>
-          <IconActivity className="text-muted-foreground" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">{acceptRate}%</div>
-        </CardContent>
-      </Card>
-
-      <Card className="flex flex-1 flex-col justify-between self-stretch bg-primary-foreground/30 lg:block">
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">
-            Total Completions
-          </CardTitle>
-          <IconCode className="text-muted-foreground" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">
-            {numeral(totalViews).format('0,0')}
-          </div>
-        </CardContent>
-      </Card>
-
-      <Card className="flex flex-1 flex-col justify-between self-stretch bg-primary-foreground/30 lg:block">
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-          <CardTitle className="text-sm font-medium">
-            Total Acceptances
-          </CardTitle>
-          <IconCheck className="h-4 w-4 text-muted-foreground" />
-        </CardHeader>
-        <CardContent>
-          <div className="text-2xl font-bold">{totalAcceptances}</div>
-        </CardContent>
-      </Card>
-    </div>
-  )
-}
+import { CompletionDailyActivity } from './completion-daily-activity'
+import { KEY_SELECT_ALL } from './constants'
 
 export function Report() {
   const searchParams = useSearchParams()
-  const [members] = useAllMembers()
+  const [members, fetchingMembers] = useAllMembers()
   const isDemoMode = useIsDemoMode()
-  const [dateRange, setDateRange] = useState<DateRange>({
-    from: moment().add(parseInt(DEFAULT_DATE_RANGE, 10), 'day').toDate(),
-    to: moment().toDate()
-  })
   const [selectedMember, setSelectedMember] = useState(KEY_SELECT_ALL)
-  const [selectedLanguage, setSelectedLanguage] = useState<Language[]>([])
-
   const sample = isDemoMode || searchParams.get('sample') === 'true'
 
-  // Query stats of selected date range
-  const [{ data: dailyStatsData, fetching: fetchingDailyState }] = useQuery({
-    query: queryDailyStats,
-    variables: {
-      start: moment(dateRange.from).startOf('day').utc().format(),
-      end: moment(dateRange.to).endOf('day').utc().format(),
-      users: selectedMember === KEY_SELECT_ALL ? undefined : [selectedMember]
-    }
-  })
-  let dailyStats: DailyStatsQuery['dailyStats'] | undefined
-  if (sample) {
-    const daysBetweenRange = eachDayOfInterval({
-      start: dateRange.from!,
-      end: dateRange.to || dateRange.from!
-    })
-    dailyStats = daysBetweenRange.map(date => {
-      const languages = [Language.Typescript, Language.Python, Language.Rust]
-      const rng = seedrandom(
-        moment(date).format('YYYY-MM-DD') + selectedMember + selectedLanguage
-      )
-      const selects = Math.ceil(rng() * 20)
-      const completions = Math.ceil(selects / 0.35)
-      return {
-        start: moment(date).utc().format(),
-        end: moment(date).add(1, 'day').utc().format(),
-        completions,
-        selects,
-        views: completions,
-        language: languages[selects % languages.length]
+  // query chat yearly stats
+  const [{ data: chatYearlyStatsData, fetching: fetchingChatYearlyStats }] =
+    useQuery({
+      query: chatDailyStatsInPastYearQuery,
+      variables: {
+        users: selectedMember === KEY_SELECT_ALL ? undefined : selectedMember
       }
     })
-  } else {
-    dailyStats = dailyStatsData?.dailyStats.map(item => ({
-      start: item.start,
-      end: item.end,
-      completions: item.completions,
-      selects: item.selects,
-      views: item.views,
-      language: item.language
-    }))
-  }
-  dailyStats = dailyStats?.filter(stats => {
-    if (selectedLanguage.length === 0) return true
-    return selectedLanguage.includes(stats.language)
-  })
 
   // Query yearly stats
   const [{ data: yearlyStatsData, fetching: fetchingYearlyStats }] = useQuery({
-    query: queryDailyStatsInPastYear,
+    query: dailyStatsInPastYearQuery,
     variables: {
       users: selectedMember === KEY_SELECT_ALL ? undefined : selectedMember
     }
   })
-  let yearlyStats: DailyStatsInPastYearQuery['dailyStatsInPastYear'] | undefined
+  let yearlyStats:
+    | Array<
+        | ArrayElementType<DailyStatsInPastYearQuery['dailyStatsInPastYear']>
+        | ArrayElementType<
+            ChatDailyStatsInPastYearQuery['chatDailyStatsInPastYear']
+          >
+      >
+    | undefined
   if (sample) {
     const daysBetweenRange = eachDayOfInterval({
       start: moment().toDate(),
       end: moment().subtract(365, 'days').toDate()
     })
     yearlyStats = daysBetweenRange.map(date => {
-      const rng = seedrandom(
-        moment(date).format('YYYY-MM-DD') + selectedMember + selectedLanguage
-      )
+      const rng = seedrandom(moment(date).format('YYYY-MM-DD') + selectedMember)
       const selects = Math.ceil(rng() * 20)
       const completions = selects + Math.floor(rng() * 10)
       return {
+        __typename: 'CompletionStats',
         start: moment(date).format('YYYY-MM-DD[T]HH:mm:ss[Z]'),
         end: moment(date).add(1, 'day').format('YYYY-MM-DD[T]HH:mm:ss[Z]'),
         completions,
@@ -202,13 +86,10 @@ export function Report() {
       }
     })
   } else {
-    yearlyStats = yearlyStatsData?.dailyStatsInPastYear.map(item => ({
-      start: item.start,
-      end: item.end,
-      completions: item.completions,
-      selects: item.selects,
-      views: item.views
-    }))
+    yearlyStats = [
+      ...(yearlyStatsData?.dailyStatsInPastYear || []),
+      ...(chatYearlyStatsData?.chatDailyStatsInPastYear || [])
+    ]
   }
 
   return (
@@ -219,7 +100,7 @@ export function Report() {
         </SubHeader>
 
         <LoadingWrapper
-          loading={fetchingDailyState}
+          loading={fetchingMembers}
           fallback={<Skeleton className="h-6 w-32" />}
         >
           <Select
@@ -248,6 +129,7 @@ export function Report() {
         </LoadingWrapper>
       </div>
 
+      {/* Yearly */}
       <LoadingWrapper
         loading={fetchingYearlyStats}
         fallback={<Skeleton className="mb-8 h-48" />}
@@ -260,168 +142,14 @@ export function Report() {
         </div>
       </LoadingWrapper>
 
-      <LoadingWrapper
-        loading={fetchingDailyState}
-        fallback={
-          <div className="flex flex-col gap-5">
-            <div className="flex justify-between gap-5">
-              <Skeleton className="h-32 flex-1" />
-              <Skeleton className="h-32 flex-1" />
-              <Skeleton className="h-32 flex-1" />
-            </div>
-            <Skeleton className="h-56" />
-          </div>
-        }
-      >
-        {/* completions */}
-        <div className="mb-10 flex flex-col gap-y-5">
-          <h1 className="text-xl font-semibold">Usage</h1>
-          <div className="flex flex-col gap-y-5">
-            <div className="-mb-2 flex flex-col items-center justify-between gap-y-1 md:flex-row md:gap-y-0">
-              <h2 className="font-semibold">Completions</h2>
-              <div className="flex flex-col gap-3 md:flex-row md:items-center md:gap-y-0">
-                <Popover>
-                  <PopoverTrigger asChild>
-                    <div className="flex h-9 w-[240px] items-center justify-between rounded-md border border-input bg-transparent px-3 py-2 text-sm shadow ring-offset-background placeholder:text-muted-foreground focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:cursor-not-allowed md:w-[150px]">
-                      <div className="w-full">
-                        {selectedLanguage.length === 0 && (
-                          <p className="w-full overflow-hidden text-ellipsis">
-                            All languages
-                          </p>
-                        )}
-                        {selectedLanguage.length === 1 && (
-                          <p className="w-full overflow-hidden text-ellipsis">
-                            {getLanguageDisplayName(selectedLanguage[0])}
-                          </p>
-                        )}
-                        {selectedLanguage.length > 1 && (
-                          <span className="px-1">
-                            {selectedLanguage.length} selected
-                          </span>
-                        )}
-                      </div>
-                      <IconChevronUpDown className="h-3 w-3" />
-                    </div>
-                  </PopoverTrigger>
-                  <PopoverContent
-                    className="w-[240px] p-0 md:w-[180px]"
-                    align="end"
-                  >
-                    <Command>
-                      <CommandInput placeholder="Language" />
-                      <CommandList>
-                        <CommandEmpty>No results found.</CommandEmpty>
+      {/* completions */}
+      <CompletionDailyActivity
+        sample={sample}
+        selectedMember={selectedMember}
+      />
 
-                        <CommandGroup>
-                          {Object.entries(Language)
-                            .sort((_, b) => (b[1] === Language.Other ? -1 : 0))
-                            .map(([_, value]) => {
-                              const isSelected =
-                                selectedLanguage.includes(value)
-                              return (
-                                <CommandItem
-                                  key={value}
-                                  onSelect={() => {
-                                    const newSelect = [...selectedLanguage]
-                                    if (isSelected) {
-                                      const idx = newSelect.findIndex(
-                                        item => item === value
-                                      )
-                                      if (idx !== -1) newSelect.splice(idx, 1)
-                                    } else {
-                                      newSelect.push(value)
-                                    }
-                                    setSelectedLanguage(newSelect)
-                                  }}
-                                  className="!pointer-events-auto cursor-pointer !opacity-100"
-                                >
-                                  <div
-                                    className={cn(
-                                      'mr-2 flex h-4 w-4 items-center justify-center rounded-sm border border-primary',
-                                      isSelected
-                                        ? 'bg-primary text-primary-foreground'
-                                        : 'opacity-50 [&_svg]:invisible'
-                                    )}
-                                  >
-                                    <IconCheck className={cn('h-4 w-4')} />
-                                  </div>
-                                  <span>{getLanguageDisplayName(value)}</span>
-                                </CommandItem>
-                              )
-                            })}
-                        </CommandGroup>
-                        {selectedLanguage.length > 0 && (
-                          <>
-                            <CommandSeparator />
-                            <CommandGroup>
-                              <CommandItem
-                                onSelect={() => setSelectedLanguage([])}
-                                className="!pointer-events-auto cursor-pointer justify-center text-center !opacity-100"
-                              >
-                                Clear filters
-                              </CommandItem>
-                            </CommandGroup>
-                          </>
-                        )}
-                      </CommandList>
-                    </Command>
-                  </PopoverContent>
-                </Popover>
-
-                <DateRangePicker
-                  options={[
-                    { label: 'Last 7 days', value: '-7d' },
-                    { label: 'Last 14 days', value: '-14d' },
-                    { label: 'Last 30 days', value: '-30d' }
-                  ]}
-                  defaultValue={DEFAULT_DATE_RANGE}
-                  onSelect={setDateRange}
-                  hasToday
-                  hasYesterday
-                />
-              </div>
-            </div>
-            <StatsSummary dailyStats={dailyStats} />
-            <DailyActivity dailyStats={dailyStats} dateRange={dateRange} />
-          </div>
-        </div>
-
-        {/* chats */}
-        <div className="mb-10 flex flex-col gap-y-5">
-          <div className="flex flex-col items-center justify-between gap-y-1 md:flex-row md:gap-y-0">
-            <h2 className="font-semibold">Chats</h2>
-            <div className="flex flex-col gap-3 md:flex-row md:items-center md:gap-y-0">
-              <DateRangePicker
-                options={[
-                  { label: 'Last 7 days', value: '-7d' },
-                  { label: 'Last 14 days', value: '-14d' },
-                  { label: 'Last 30 days', value: '-30d' }
-                ]}
-                defaultValue={DEFAULT_DATE_RANGE}
-                onSelect={setDateRange}
-                hasToday
-                hasYesterday
-              />
-            </div>
-          </div>
-          <div className="grid md:grid-cols-3">
-            <Card className="flex flex-1 flex-col justify-between self-stretch bg-primary-foreground/30 lg:block">
-              <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-2">
-                <CardTitle className="text-sm font-medium">
-                  Total Chats
-                </CardTitle>
-                <IconMessageSquare className="text-muted-foreground" />
-              </CardHeader>
-              <CardContent>
-                <div className="text-2xl font-bold">
-                  {numeral(1265).format('0,0')}
-                </div>
-              </CardContent>
-            </Card>
-          </div>
-          <ChatDailyActivity dailyStats={dailyStats} dateRange={dateRange} />
-        </div>
-      </LoadingWrapper>
+      {/* chats */}
+      <ChatDailyActivity sample={sample} selectedMember={selectedMember} />
     </div>
   )
 }
